refactor: remove unused Perps deposit flows and components
- Remove unused PERPS.DEPOSIT, POSITION_DETAILS, ORDER_HISTORY, ORDER_DETAILS routes
- Delete PerpsTokenSelector component (unused)
- Remove old deposit hooks (usePerpsDeposit, usePerpsDepositQuote)
- Move PerpsToken type to proper types file for withdrawal hooks
- Remove PERPS_NAVIGATION_ARCHITECTURE.md documentation file
- Keep PerpsTransactionsView and transaction detail views (actively used)
- Preserve all confirmations folder content (new deposit flow)
