--- app/components/UI/Perps/components/PerpsTokenSelector/index.ts
@@ -1,2 +1,0 @@
-export { default } from './PerpsTokenSelector';
-export type { PerpsToken } from './PerpsTokenSelector';
