+++ app/components/UI/Perps/hooks/index.ts
@@ -1,6 +1,4 @@
 // Core hooks (direct controller access)
-export { usePerpsDeposit } from './usePerpsDeposit';
-export { usePerpsDepositQuote } from './usePerpsDepositQuote';
 export { usePerpsMarkets } from './usePerpsMarkets';
 export { usePerpsNetwork } from './usePerpsNetwork';
 export { usePerpsNetworkConfig } from './usePerpsNetworkConfig';
@@ -66,7 +64,6 @@ export { useStableArray } from './useStableArray';
 // Stream hooks (WebSocket subscriptions)
 export * from './stream';
 
-// - usePerpsDeposit (deposit flows)
 // - usePerpsPendingOrders (order management)
 // - usePerpsError (error handling)
 // - usePerpsLivePositions (live position tracking)
