--- app/components/UI/Perps/hooks/usePerpsDeposit.ts
@@ -1,10 +1,0 @@
-import { useSelector } from 'react-redux';
-import { selectPerpsDepositState } from '../selectors/perpsController';
-
-/**
- * Consolidated hook for deposit state
- */
-// TODO: Replace hook with selector in components.
-export function usePerpsDeposit() {
-  return useSelector(selectPerpsDepositState);
-}
