+++ app/components/UI/Perps/hooks/usePerpsPaymentTokens.ts
@@ -16,7 +16,7 @@ import {
   TRADING_DEFAULTS,
   USDC_ARBITRUM_MAINNET_ADDRESS,
 } from '../constants/hyperLiquidConfig';
-import type { PerpsToken } from '../components/PerpsTokenSelector';
+import type { PerpsToken } from '../types';
 
 /**
  * Hook to get all payment tokens for Perps, including:
