+++ app/components/UI/Perps/hooks/useWithdrawTokens.ts
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux';
 
 import { selectIsIpfsGatewayEnabled } from '../../../../selectors/preferencesController';
 import { selectTokenList } from '../../../../selectors/tokenListController';
-import type { PerpsToken } from '../components/PerpsTokenSelector';
+import type { PerpsToken } from '../types';
 import {
   ARBITRUM_MAINNET_CHAIN_ID,
   HYPERLIQUID_MAINNET_CHAIN_ID,
