+++ app/components/UI/Perps/types/index.ts
@@ -61,6 +61,7 @@ export interface CandleData {
 
 // Export all configuration types directly
 export * from './config';
+export * from './token';
 
 /**
  * Standard Structure for Perps LaunchDarkly Feature Flags
