+++ app/components/UI/Perps/types/token.ts
@@ -1,0 +1,7 @@
+import type { BridgeToken } from '../../Bridge/types';
+
+/**
+ * PerpsToken is an alias for BridgeToken for backward compatibility
+ * This type represents tokens that can be used in Perps trading
+ */
+export type PerpsToken = BridgeToken;
