+++ app/constants/navigation/Routes.ts
@@ -213,20 +213,16 @@ const Routes = {
     ROOT: 'Perps',
     TRADING_VIEW: 'PerpsTradingView',
     ORDER: 'PerpsOrder',
-    DEPOSIT: 'PerpsDeposit',
     WITHDRAW: 'PerpsWithdraw',
     POSITIONS: 'PerpsPositions',
     MARKETS: 'PerpsMarketListView',
     MARKET_DETAILS: 'PerpsMarketDetails',
-    POSITION_DETAILS: 'PerpsPositionDetails',
     TUTORIAL: 'PerpsTutorial',
     MODALS: {
       ROOT: 'PerpsModals',
       QUOTE_EXPIRED_MODAL: 'PerpsQuoteExpiredModal',
       BALANCE_MODAL: 'PerpsBalanceModal',
     },
-    ORDER_HISTORY: 'PerpsOrderHistory',
-    ORDER_DETAILS: 'PerpsOrderDetails',
     POSITION_TRANSACTION: 'PerpsPositionTransaction',
     ORDER_TRANSACTION: 'PerpsOrderTransaction',
     FUNDING_TRANSACTION: 'PerpsFundingTransaction',
